+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 #### v1.1.0 > v1.2.0
 
+-   Add video rotation filter ([#347](https://github.com/datarhei/restreamer/discussions/347))
+-   Add video h/v flip filter
+-   Add audio volume filter ([#313](https://github.com/datarhei/restreamer/issues/313))
+-   Add audio loudness normalization filter
 -   Add HLS Master playlist (requires FFmpeg hlsbitrate.patch) (thx Dwaynarang, Electra Player compatibility)
 -   Add linkedIn & Azure Media Services to publication services (thx kalashnikov)
 -   Add AirPlay support with silvermine videojs plugin
@@ -14,6 +18,7 @@
 -   Add Polish translations (thx Robert Rykała)
 -   Mod extends the datarhei Core publication service with srt streaming
 -   Mod Allow decoders and encoders to set global options
+-   Fix player problem with different stream formats (9:16)
 -   Mod Allow trailing slash on Core address
 -   Fix process report naming
 -   Fix publication service icon styles